Rename navbar to masterbar This aligns better with the source component (client/layout/masterbar)
